+++ packages/integration-sdk-private-test-utils/__fixtures__/docsInstanceEntitiesAlphabetize/src/index.ts
@@ -1,0 +1,10 @@
+import fetchDataSteps from './steps/fetchData';
+import {
+  IntegrationInvocationConfig,
+  IntegrationInstanceConfig,
+} from '@jupiterone/integration-sdk-core';
+
+export const invocationConfig: IntegrationInvocationConfig<IntegrationInstanceConfig> = {
+  instanceConfigFields: {},
+  integrationSteps: [fetchDataSteps],
+};