+++ packages/integration-sdk-private-test-utils/__fixtures__/docsInstanceRelationshipsAlphabetize/src/index.ts
@@ -1,0 +1,10 @@
+import fetchAccountsStep from './steps/fetchAccounts';
+import {
+  IntegrationInvocationConfig,
+  IntegrationInstanceConfig,
+} from '@jupiterone/integration-sdk-core';
+
+export const invocationConfig: IntegrationInvocationConfig<IntegrationInstanceConfig> = {
+  instanceConfigFields: {},
+  integrationSteps: [fetchAccountsStep],
+};